ast: Refactor utils
