+++ common/ast/src/parser/expr.rs
@@ -25,11 +25,14 @@ use pratt::PrattParser;
 use pratt::Precedence;
 
 use crate::ast::*;
+use crate::input::Input;
+use crate::input::WithSpan;
+use crate::match_token;
 use crate::parser::query::*;
 use crate::parser::token::*;
 use crate::parser::unescape::unescape;
-use crate::parser::util::*;
 use crate::rule;
+use crate::util::*;
 use crate::Error;
 use crate::ErrorKind;
