+++ common/ast/src/rule.rs
@@ -1,4 +1,4 @@
-// Copyright 2021 Datafuse Labs.
+// Copyright 2022 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,19 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod error;
-pub use error::Backtrace;
-pub use error::DisplayError;
-pub use error::Error;
-pub use error::ErrorKind;
+use nom::Slice;
 
-mod input;
-pub use input::Input;
-
-mod util;
-pub use util::match_text;
-pub use util::match_token;
-
-pub mod ast;
-pub mod parser;
-pub mod udfs;
+use crate::parser::token::Token;
+use crate::parser::token::TokenKind;
+use crate::util::IResult;
+use crate::Error;
+use crate::ErrorKind;
+use crate::Input;
