+++ common/ast/tests/it/parser.rs
@@ -19,10 +19,10 @@ use common_ast::parser::parse_sql;
 use common_ast::parser::query::*;
 use common_ast::parser::token::*;
 use common_ast::parser::tokenize_sql;
-use common_ast::parser::util::Input;
 use common_ast::rule;
 use common_ast::Backtrace;
 use common_ast::DisplayError;
+use common_ast::Input;
 use common_exception::Result;
 use goldenfile::Mint;
 use nom::Parser;
